1.8.0 DEV_21
변장술 이름,스킨 변경 설명 추가
복수 반사대미지 소숫점단위 수정 가능
엑시즈 임시 해결

이야 21까지 가게될줄은...

+++ Addon/src/com/cokes86/cokesaddon/ability/list/Revenge.java
@@ -24,7 +24,7 @@ import java.text.DecimalFormat;
 		"상대방을 공격할 시 최근에 플레이어에게 받았던 대미지의 $[PERCENTAGE]% 만큼의 고정 마법 대미지를 상대방에게 추가적으로 입힙니다."
 })
 public class Revenge extends CokesAbility {
-	public static final Config<Integer> PERCENTAGE = Config.of(Revenge.class, "반사대미지(%)", 40, FunctionalInterfaceUnit.positive());
+	public static final Config<Double> PERCENTAGE = Config.of(Revenge.class, "반사대미지(%)", 40d, FunctionalInterfaceUnit.positive());
 	private final DecimalFormat df = new DecimalFormat("0.##");
 	private double finalDamage = 0;
 	private final ActionbarChannel ac = newActionbarChannel();
@@ -35,7 +35,7 @@ public class Revenge extends CokesAbility {
 
 	public void onUpdate(Update update) {
 		if (update == Update.RESTRICTION_CLEAR) {
-			ac.update(ChatColor.BLUE + "반사고정대미지 : " + df.format(finalDamage * PERCENTAGE.getValue() / (double) 100));
+			ac.update(ChatColor.BLUE + "반사고정대미지 : " + df.format(finalDamage * PERCENTAGE.getValue() / 100));
 		}
 	}
